Set UID to the current UID
Up to this point the UID of the `nodejs` user has been has been
automatically set to 998. In order to avoid a breaking change and to
clearly document the UID for downstream we should explicitly set the UID
to 998.

This could be changed when we move to the next major version of Node.js
or should we move to a more sophisticated way of versioning this image.