Use latest npm
npm@6 allows for use of `npm ci` command which install _strictly_ the dependencies as per the package-lock in a project and will fail if the package-lock and package.json are inconsistent. 

This provides enhanced security for deployments as it gives precise control over the exact versions of the full dependency tree.